refactor(tests): split test_wizard.py into two files
